# tests/golden/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VoightKampff OCR API",
    "version": "0.3.0"
  },
  "paths": {
    "/v1/ocr/passport": {
      "post": {
        "tags": [
          "v1 OCR"
        ],
        "summary": "Passport",
        "operationId": "passport_v1_ocr_passport_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_passport_v1_ocr_passport_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OcrResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/ocr/id-card": {
      "post": {
        "tags": [
          "v1 OCR"
        ],
        "summary": "Id Card",
        "operationId": "id_card_v1_ocr_id_card_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_id_card_v1_ocr_id_card_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OcrResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/ocr/driving-license": {
      "post": {
        "tags": [
          "v1 OCR"
        ],
        "summary": "Driving License",
        "operationId": "driving_license_v1_ocr_driving_license_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_driving_license_v1_ocr_driving_license_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OcrResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/ocr/passport/batch": {
      "post": {
        "tags": [
          "v1 OCR"
        ],
        "summary": "Passport Batch",
        "operationId": "passport_batch_v1_ocr_passport_batch_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_passport_batch_v1_ocr_passport_batch_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OcrBatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/ocr/id-card/batch": {
      "post": {
        "tags": [
          "v1 OCR"
        ],
        "summary": "Id Card Batch",
        "operationId": "id_card_batch_v1_ocr_id_card_batch_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_id_card_batch_v1_ocr_id_card_batch_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OcrBatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/ocr/driving-license/batch": {
      "post": {
        "tags": [
          "v1 OCR"
        ],
        "summary": "Driving License Batch",
        "operationId": "driving_license_batch_v1_ocr_driving_license_batch_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_driving_license_batch_v1_ocr_driving_license_batch_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OcrBatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BatchItemResult": {
        "properties": {
          "index": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Index"
          },
          "input": {
            "$ref": "#/components/schemas/DocumentInput"
          },
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DocumentResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorResult"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "index",
          "input",
          "success"
        ],
        "title": "BatchItemResult"
      },
...
      "Body_driving_license_batch_v1_ocr_driving_license_batch_post": {
        "properties": {
          "images": {
            "items": {
              "type": "string",
              "format": "binary",
              "contentMediaType": "application/octet-stream"
            },
            "type": "array",
            "title": "Images",
            "description": "Select one or more image files, ZIP archives, or a mixture of both. ZIP archives may contain images in nested folders.",
            "default": []
          },
          "archive": {
            "anyOf": [
              {
                "type": "string",
                "format": "binary",
                "contentMediaType": "application/octet-stream"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archive"
          }
        },
        "type": "object",
        "title": "Body_driving_license_batch_v1_ocr_driving_license_batch_post"
      },
      "Body_driving_license_v1_ocr_driving_license_post": {
        "properties": {
          "image": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Image"
          }
        },
        "type": "object",
        "required": [
          "image"
        ],
        "title": "Body_driving_license_v1_ocr_driving_license_post"
      },
      "Body_id_card_batch_v1_ocr_id_card_batch_post": {
        "properties": {
          "archive": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Archive"
          }
        },
        "type": "object",
        "required": [
          "archive"
        ],
        "title": "Body_id_card_batch_v1_ocr_id_card_batch_post"
      },
...
      "Body_id_card_v1_ocr_id_card_post": {
        "properties": {
          "front": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Front"
          },
          "back": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Back"
          }
        },
        "type": "object",
        "required": [
          "front",
          "back"
        ],
        "title": "Body_id_card_v1_ocr_id_card_post"
      },
...
      "Body_passport_batch_v1_ocr_passport_batch_post": {
        "properties": {
          "images": {
            "items": {
              "type": "string",
              "format": "binary",
              "contentMediaType": "application/octet-stream"
            },
            "type": "array",
            "title": "Images",
            "description": "Select one or more image files, ZIP archives, or a mixture of both. ZIP archives may contain images in nested folders.",
            "default": []
          },
          "archive": {
            "anyOf": [
              {
                "type": "string",
                "format": "binary",
                "contentMediaType": "application/octet-stream"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archive"
          }
        },
        "type": "object",
        "title": "Body_passport_batch_v1_ocr_passport_batch_post"
      },
...
      "Body_passport_v1_ocr_passport_post": {
        "properties": {
          "image": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Image"
          }
        },
        "type": "object",
        "required": [
          "image"
        ],
        "title": "Body_passport_v1_ocr_passport_post"
      },
...
      "BoundingBox": {
        "properties": {
          "x1": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "X1"
          },
          "y1": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Y1"
          },
          "x2": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "X2"
          },
          "y2": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Y2"
          }
        },
        "type": "object",
        "required": [
          "x1",
          "y1",
          "x2",
          "y2"
        ],
        "title": "BoundingBox"
      },
      "Confidence": {
        "properties": {
          "score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Score"
          },
          "source": {
            "$ref": "#/components/schemas/ConfidenceSource"
          },
          "calibrated_probability": {
            "type": "boolean",
            "title": "Calibrated Probability",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "score",
          "source"
        ],
        "title": "Confidence"
      },
      "ConfidenceSource": {
        "type": "string",
        "enum": [
          "document_detection",
          "ocr_token",
          "ocr_token_mean",
          "ocr_token_minimum"
        ],
        "title": "ConfidenceSource"
      },
      "DocumentInput": {
        "properties": {
          "document_type": {
            "$ref": "#/components/schemas/DocumentType"
          },
          "image": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image"
          },
          "front": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Front"
          },
          "back": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Back"
          }
        },
        "type": "object",
        "required": [
          "document_type"
        ],
        "title": "DocumentInput",
        "description": "Logical upload names; the HTTP transport supplies the corresponding files."
      },
      "DocumentResult": {
        "properties": {
          "document_type": {
            "$ref": "#/components/schemas/DocumentType"
          },
          "layout": {
            "type": "string",
            "title": "Layout"
          },
          "fields": {
            "additionalProperties": {
              "$ref": "#/components/schemas/FieldResult"
            },
            "type": "object",
            "title": "Fields"
          },
          "document_confidence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Confidence"
              },
              {
                "type": "null"
              }
            ]
          },
          "mrz": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MrzResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "validations": {
            "items": {
              "$ref": "#/components/schemas/ValidationResult"
            },
            "type": "array",
            "title": "Validations"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "timings": {
            "$ref": "#/components/schemas/TimingResult"
          }
        },
        "type": "object",
        "required": [
          "document_type",
          "layout",
          "fields",
          "timings"
        ],
        "title": "DocumentResult"
      },
      "DocumentType": {
        "type": "string",
        "enum": [
          "passport",
          "id_card",
          "driving_license"
        ],
        "title": "DocumentType"
      },
      "ErrorCode": {
        "type": "string",
        "enum": [
          "invalid_upload",
          "unsupported_media_type",
          "upload_too_large",
          "too_many_items",
          "invalid_archive",
          "missing_document_side",
          "invalid_document",
          "profile_unavailable",
          "model_unavailable",
          "queue_full",
          "resource_exhausted",
          "processing_failed"
        ],
        "title": "ErrorCode"
      },
      "ErrorResult": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "code",
          "detail"
        ],
        "title": "ErrorResult"
      },
      "FieldResult": {
        "properties": {
          "value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          },
          "raw_text": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Raw Text"
          },
          "confidence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Confidence"
              },
              {
                "type": "null"
              }
            ]
          },
          "bounding_box": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BoundingBox"
              },
              {
                "type": "null"
              }
            ]
          },
          "region": {
            "type": "string",
            "title": "Region"
          }
        },
        "type": "object",
        "required": [
          "value",
          "region"
        ],
        "title": "FieldResult"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MrzResult": {
        "properties": {
          "raw_lines": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Raw Lines"
          },
          "fields": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "title": "Fields"
          },
          "validations": {
            "items": {
              "$ref": "#/components/schemas/ValidationResult"
            },
            "type": "array",
            "title": "Validations"
          }
        },
        "type": "object",
        "title": "MrzResult"
      },
      "OcrBatchResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          },
          "succeeded": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Succeeded"
          },
          "failed": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Failed"
          },
          "total_seconds": {
            "type": "number",
            "minimum": 0.0,
            "title": "Total Seconds"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/BatchItemResult"
            },
            "type": "array",
            "title": "Items"
          },
          "diagnostics": {
            "additionalProperties": true,
            "type": "object",
            "title": "Diagnostics"
          }
        },
        "type": "object",
        "required": [
          "total",
          "succeeded",
          "failed",
          "total_seconds",
          "items"
        ],
        "title": "OcrBatchResponse"
      },
      "OcrResponse": {
        "properties": {
          "result": {
            "$ref": "#/components/schemas/DocumentResult"
          }
        },
        "type": "object",
        "required": [
          "result"
        ],
        "title": "OcrResponse"
      },
      "TimingResult": {
        "properties": {
          "total_seconds": {
            "type": "number",
            "minimum": 0.0,
            "title": "Total Seconds"
          },
          "stages": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Stages"
          }
        },
        "type": "object",
        "required": [
          "total_seconds"
        ],
        "title": "TimingResult"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "ValidationResult": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "status": {
            "$ref": "#/components/schemas/ValidationStatus"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Fields"
          }
        },
        "type": "object",
        "required": [
          "code",
          "status"
        ],
        "title": "ValidationResult"
      },
      "ValidationStatus": {
        "type": "string",
        "enum": [
          "passed",
          "failed",
          "warning",
          "not_run"
        ],
        "title": "ValidationStatus"
      },
...
    }
  },
...
}